migration-tool/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
rapidfuzz==3.10.0
python-dotenv==1.0.1
python-multipart==0.0.12
